restructuring folders and verification
