+++ ignite-generator/src/screen/index.es
@@ -1,14 +1,12 @@
 #! /usr/bin/env node
 'use strict'
 
-import Generator from '../Generator'
 import { NamedBase } from 'yeoman-generator'
 
 class ScreenGenerator extends NamedBase {
 
   generateFile () {
-    console.log('generate screen - ' + this.name)
-    Generator.hydrateComponent('Containers', this.name)
+    console.log('Coming soon generate screen - ' + this.name)
   }
 }
